+++ src/Phan/Language/Type/BoolType.php
@@ -47,7 +47,9 @@ final class BoolType extends ScalarType
 
     public function canCastToDeclaredType(CodeBase $code_base, Context $context, Type $other): bool
     {
-        return $other->isInBoolFamily() || (!$context->isStrictTypes() && parent::canCastToDeclaredType($code_base, $context, $other));
+        return $other->isInBoolFamily() ||
+            $other instanceof MixedType ||
+            (!$context->isStrictTypes() && parent::canCastToDeclaredType($code_base, $context, $other));
     }
 
     public function isPossiblyTrue(): bool