+++ src/Phan/Language/Type/TrueType.php
@@ -91,7 +91,9 @@ final class TrueType extends ScalarType
 
     public function canCastToDeclaredType(CodeBase $code_base, Context $context, Type $other): bool
     {
-        return $other->isInBoolFamily() || (!$context->isStrictTypes() && parent::canCastToDeclaredType($code_base, $context, $other));
+        return $other->isInBoolFamily() ||
+            $other instanceof MixedType ||
+            (!$context->isStrictTypes() && parent::canCastToDeclaredType($code_base, $context, $other));
     }
 
     // public function getTypeAfterIncOrDec() : UnionType - doesn't need to be changed